+++ CHANGELOG.md
@@ -128,6 +128,8 @@ This project uses a [semantic versioning](https://semver.org/) scheme as the bas
 
 #### Deprecated
 - `CLI` toolset marked as deprecated and will be removed from package in version `3.0.0` release. Deprecation notification will be displayed on each `cli` call.
+- `.core.BaseLang.is_file_exist()` marked as deprecated and replaced with `.file_exists()` method. Will be removed in version `3.0.0`.
+  > Note that changes to `.core.BaseLang` will be reflected on all inherited from it classes *(`JSON_Format`, `YAML_Format`, ...)*.
 
 
 ### **2.2.0dev** : nightly